Move the `/pkg/assets/dynamic.swiftFS` implem into its own file
At the moment the `fs.go` file contains 3 differents implementation.

Separate them into differents files would improve the readability as
we can quickly see what implementations are available.
